+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Command-line option `--init` to create a new changelog if it does not exist.
+- Command-line option `--check` to check if changelog has any changes under \[Unreleased\].
+
+### Removed
+- **[BREAKING CHANGE]** Removed support for Node versions <= 8
 
 ## [1.1.0] - 2018-01-02
 ### Added